search: rename 'unindexed' -> 'textsearch' package
