+++ internal/search/textsearch/textsearch_test.go
@@ -1,4 +1,4 @@
-package unindexed
+package textsearch
 
 import (
 	"context"
